Add type declarations for higher-order components
