+++ assets/src/components/v2/persistent_wrapper.tsx
@@ -1,7 +1,16 @@
-import React, { useContext, useEffect, useState } from "react";
+import React, { ComponentType, useContext, useEffect, useState } from "react";
 import { LastFetchContext } from "Components/v2/screen_container";
 
-const PersistentWrapper = ({ WrappedComponent, ...data }) => {
+interface WrappedComponentProps {
+  onFinish: () => void;
+  lastUpdate: number | null;
+}
+
+interface Props {
+  WrappedComponent: ComponentType<WrappedComponentProps>;
+}
+
+const PersistentWrapper: ComponentType<Props> = ({ WrappedComponent, ...data }) => {
   const lastFetch = useContext(LastFetchContext);
 
   const [visibleData, setVisibleData] = useState(data);
@@ -34,6 +43,9 @@ const PersistentWrapper = ({ WrappedComponent, ...data }) => {
  * Call this function on a `WrappedComponent` to allow it to persist across
  * data refreshes.
  *
+ * Consider extending the `WrappedComponentProps` type exported by this module when
+ * defining your `WrappedComponent`'s props.
+ *
  * `WrappedComponent` must expect the following props:
  * - `onFinish`: A callback that `WrappedComponent` can call to indicate that it is ready to receive
  *   fresh data on the next refresh.
@@ -43,7 +55,9 @@ const PersistentWrapper = ({ WrappedComponent, ...data }) => {
  * Any other props (e.g. data to be rendered) are passed through to `WrappedComponent` unchanged.
  */
 const makePersistent =
-  (Component) =>
+  (Component: ComponentType<WrappedComponentProps>) =>
     ({ ...data }) =>
       <PersistentWrapper {...data} WrappedComponent={Component} />;
+
 export default makePersistent;
+export { WrappedComponentProps };
